chore: Restructure files, bind repository
