+++ src/Api/BasicDiscussionAttributes.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace FoF\BestAnswer;
+namespace FoF\BestAnswer\Api;
 
 use Flarum\Api\Serializer\BasicDiscussionSerializer;
 use Flarum\Discussion\Discussion;
