+++ src/Api/UserBestAnswerCount.php
@@ -9,10 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace FoF\BestAnswer;
+namespace FoF\BestAnswer\Api;
 
 use Flarum\Api\Serializer\UserSerializer;
 use Flarum\User\User;
+use FoF\BestAnswer\Repository\BestAnswerRepository;
 
 class UserBestAnswerCount
 {
